Update and rename Timer_2-1_hardcode to Timer_programmable
I added a programmable functionality to the timer. This way it accepts 2 arguments to its main function x & y, for x-minute games with y-second increments. 

It also switches display style when there are less than 10 seconds left to be seconds and tenths of seconds.

Still needs to be integrated with Ajay's display code